+++ R/preprocessing5.R
@@ -1308,16 +1308,16 @@ SCTransform.StdAssay <- function(
         if (verbose){
           message("Getting residuals for block ", i, "(of ", length(cells.grid), ") for ", dataset.names[[dataset.index]], " dataset")
         }
-        counts.vp <- as.sparse(x = layer.data[, vp])
+        counts.vp <- as.sparse(x = layer.data[, vp, drop=FALSE])
         cell.attr.object <- cell.attr.layer[colnames(x = counts.vp),, drop=FALSE]
         vst_out <- vst_out.reference
 
         vst_out$cell_attr <- cell.attr.object
-        vst_out$gene_attr <- vst_out$gene_attr[variable.features,]
+        vst_out$gene_attr <- vst_out$gene_attr[variable.features,,drop=FALSE]
         if (return.only.var.genes){
           new_residual <- get_residuals(
             vst_out = vst_out,
-            umi = counts.vp[variable.features,],
+            umi = counts.vp[variable.features,,drop=FALSE],
             residual_type = "pearson",
             min_variance = min_var,
             res_clip_range = res_clip_range,
@@ -1326,7 +1326,7 @@ SCTransform.StdAssay <- function(
         } else {
           new_residual <- get_residuals(
             vst_out = vst_out,
-            umi = counts.vp[all_features,],
+            umi = counts.vp[all_features,,drop=FALSE],
             residual_type = "pearson",
             min_variance = min_var,
             res_clip_range = res_clip_range,
@@ -1336,7 +1336,7 @@ SCTransform.StdAssay <- function(
         vst_out$y <- new_residual
         corrected_counts[[i]] <- correct_counts(
           x = vst_out,
-          umi = counts.vp[all_features,],
+          umi = counts.vp[all_features,,drop=FALSE],
           verbosity = FALSE# as.numeric(x = verbose) * 2
         )
         residuals[[i]] <- new_residual
@@ -1345,15 +1345,14 @@ SCTransform.StdAssay <- function(
       new.residuals <- Reduce(cbind, residuals)
       corrected_counts <- Reduce(cbind, corrected_counts)
       cell_attrs <- Reduce(rbind, cell_attrs)
-      vst_out.reference$cell_attr <- cell_attrs[colnames(new.residuals),]
+      vst_out.reference$cell_attr <- cell_attrs[colnames(new.residuals),,drop=FALSE]
       SCTModel.list <- PrepVSTResults(vst.res = vst_out.reference, cell.names = all_cells)
       SCTModel.list <- list(model1 = SCTModel.list)
       # scale data here as do.center and do.scale are set to FALSE inside
       new.residuals <- ScaleData(
         new.residuals,
         features = NULL,
-        #vars.to.regress = vars.to.regress,
-        #latent.data = cell.attr[, vars.to.regress, drop = FALSE],
+        vars.to.regress = vars.to.regress,
         model.use = 'linear',
         use.umi = FALSE,
         do.scale = do.scale,