+++ NEWS.md
@@ -1,3 +1,10 @@
+# Unreleased 
+
+## Changes
+
+- Fixed `SCTransform` to handle `vars.to.regress` ([#8148](https://github.com/satijalab/seurat/issues/8148))
+- Fixed `SCTransform` to handle fetching residuals ([#8185](https://github.com/satijalab/seurat/issues/8185))
+
 # Seurat 5.0.1 (2023-11-16)
 
 ## Changes